samples/cext-trey-research-python/requirements.txt
aiohttp
azure-data-tables
uuid
azure-functions
debugpy
requests==2.31.0